+++ services/QuillLMS/app/models/concerns/public_progress_reports.rb
@@ -3,6 +3,7 @@
 module PublicProgressReports
   include DiagnosticReports
   include GetScoreForQuestion
+  include EvidenceReports
 
   extend ActiveSupport::Concern
 
@@ -13,12 +14,6 @@ module PublicProgressReports
   CONNECT_OPTIMAL_FINAL_ATTEMPT_FEEDBACK = "That's a strong sentence!"
   CONNECT_SUBOPTIMAL_FINAL_ATTEMPT_SENTENCE_COMBINING_FEEDBACK = "Nice try. Let's try a multiple choice question."
   CONNECT_SUBOPTIMAL_FINAL_ATTEMPT_FILL_IN_BLANKS_FEEDBACK =  "Good try! Compare your response to the strong responses, and then go to on to the next question."
-  # this constant is also used on the frontend to generate hardcoded feedback for students to see (see services/QuillLMS/client/app/constants/evidence.ts) and should be updated there if it's updated here.
-  EVIDENCE_SUBOPTIMAL_SPELLING_OR_GRAMMAR_FINAL_ATTEMPT_FEEDBACK = "You completed four revisions! You’ve found the right piece of evidence, but there may still be spelling or grammar changes you could make to improve your sentence. Read your sentence one more time and think about what changes you could make. Then move on to the next prompt."
-  EVIDENCE_SUBOPTIMAL_FINAL_ATTEMPT_FEEDBACK = "You completed four revisions!"
-  EVIDENCE_OPTIMAL_FINAL_ATTEMPT_FEEDBACK = "Nice work!"
-  EVIDENCE_FINAL_ATTEMPT_NUMBER = 5
-  EVIDENCE_SUBOPTIMAL_SPLIT_TEXT = 'Here are'
 
   def last_completed_diagnostic
     diagnostic_activity_ids = Activity.diagnostic_activity_ids
@@ -344,49 +339,6 @@ module PublicProgressReports
     end
   end
 
-  def get_feedback_from_feedback_history(activity_session, prompt_text, attempt_number)
-    feedback_history = get_feedback_history_from_activity_session_prompt_text_and_attempt_number(activity_session, prompt_text, attempt_number)
-    feedback_history&.feedback_text
-  end
-
-  def get_feedback_history_from_activity_session_prompt_text_and_attempt_number(activity_session, prompt_text, attempt_number)
-    feedback_histories = activity_session.feedback_histories
-
-    return nil if feedback_histories.empty? || prompt_text.blank? || attempt_number.blank?
-
-    prompt = get_evidence_prompt_from_activity_and_prompt_text(activity_session, prompt_text)
-
-    feedback_histories.select {|fh| fh.attempt == attempt_number.to_i && fh.prompt_id == prompt&.id }&.first
-  end
-
-  def get_suboptimal_final_attempt_evidence_feedback_from_activity_session_and_prompt_text(activity_session, prompt_text)
-    feedback_history = get_feedback_history_from_activity_session_prompt_text_and_attempt_number(activity_session, prompt_text, EVIDENCE_FINAL_ATTEMPT_NUMBER)
-
-    if [FeedbackHistory::GRAMMAR, FeedbackHistory::RULES_BASED_THREE, FeedbackHistory::SPELLING].include?(feedback_history&.feedback_type)
-      return EVIDENCE_SUBOPTIMAL_SPELLING_OR_GRAMMAR_FINAL_ATTEMPT_FEEDBACK
-    end
-
-    prompt = get_evidence_prompt_from_activity_and_prompt_text(activity_session, prompt_text)
-
-    format_max_attempts_feedback(prompt&.max_attempts_feedback) || EVIDENCE_SUBOPTIMAL_FINAL_ATTEMPT_FEEDBACK
-  end
-
-  def format_max_attempts_feedback(max_attempts_feedback)
-    return nil unless max_attempts_feedback
-
-    # remove HTML, split to take everything before the split-off text, remove leading and trailing whitespace, and make sure any punctuation is followed by a space
-    ActionView::Base.full_sanitizer.sanitize(max_attempts_feedback)
-      .split(EVIDENCE_SUBOPTIMAL_SPLIT_TEXT)
-      .first
-      .strip
-      .gsub(/(\.|,|;|!|\?)(?=\S)/, '\1 ')
-  end
-
-  def get_evidence_prompt_from_activity_and_prompt_text(activity_session, prompt_text)
-    prompt_ids = Evidence::Activity.find_by(parent_activity_id: activity_session.activity_id)&.prompt_ids
-    Evidence::Prompt.where(id: prompt_ids, text: prompt_text)&.first
-  end
-
   # rubocop:disable Metrics/CyclomaticComplexity
   def generate_recommendations_for_classroom(current_user, unit_id, classroom_id, activity_id)
     set_activity_sessions_and_assigned_students_for_activity_classroom_and_unit(activity_id, classroom_id, unit_id)
@@ -525,14 +477,6 @@ module PublicProgressReports
       })
     end
     question_array
-  end
-
-  private def evidence_final_attempt_feedback(activity_session, score, prompt_text, attempt_number)
-    if score == 0 && attempt_number == EVIDENCE_FINAL_ATTEMPT_NUMBER
-      get_suboptimal_final_attempt_evidence_feedback_from_activity_session_and_prompt_text(activity_session, prompt_text)
-    else
-      get_feedback_from_feedback_history(activity_session, prompt_text, attempt_number) || EVIDENCE_OPTIMAL_FINAL_ATTEMPT_FEEDBACK
-    end
   end
 
   private def grammar_final_attempt_feedback(score)
